ner_image_classification/requirements.txt
Pillow==12.1.1
scikit_learn==1.8.0
torch==2.10.0
torchvision==0.25.0
transformers==5.3.0
numpy
